Refactor Data Docs Link Class Again
Turns out this does work as a class in the operator class file,
and can be referenced by the operator just fine, but the egg or
dist file for the package has to be present in site-packages for
this to get loaded properly if it's not a plugin.
